Removing the adapter and renaming the trigger to SqsEventHandler
