--- dotnet-test-samples/sqs-lambda/src/SQSEventHandler/Adapters/ISqsEventAdapter.cs
@@ -1,11 +1,0 @@
-using System.Collections.Generic;
-using Amazon.Lambda.SQSEvents;
-
-namespace SqsEventHandler.Adapters;
-
-public interface ISqsEventAdapter
-{
-    SQSBatchResponse GetSqsBatchResponse();
-    List<SQSEvent.SQSMessage> GetSqsBatchMessages();
-    void SetSqsBatchItemFailures(List<SQSBatchResponse.BatchItemFailure> batchItemFailures);
-}
