+++ dotnet-test-samples/sqs-lambda/src/SQSEventHandler/Functions/ProcessEmployeeFunction.cs
@@ -1,19 +1,19 @@
 using System;
 using System.Threading.Tasks;
 using Amazon.Lambda.Core;
+using SqsEventHandler.Handlers;
 using SqsEventHandler.Models;
-using SqsEventHandler.Triggers;
 
 // Assembly attribute to enable the Lambda function's JSON input to be converted into a .NET class.
 [assembly: LambdaSerializer(typeof(Amazon.Lambda.Serialization.SystemTextJson.DefaultLambdaJsonSerializer))]
 
-namespace SqsEventHandler;
+namespace SqsEventHandler.Functions;
 
 /// <summary>
 /// This class implements the business logic of the function. The function handler can be found in
-/// the base abstract class SqsEventTrigger::Handler
+/// the base abstract class SqsEventHandler::Handler
 /// </summary>
-public class ProcessEmployeeFunction : SqsEventTrigger<Employee>
+public class ProcessEmployeeFunction : SqsEventHandler<Employee>
 {
     public override async Task ProcessSqsMessage(Employee message, ILambdaContext lambdaContext)
     {
