+++ dotnet-test-samples/sqs-lambda/src/SQSEventHandler/Models/Employee.cs
@@ -5,13 +5,17 @@ namespace SqsEventHandler.Models;
 
 public record Employee
 {
-    [JsonProperty("employee_id")] public string EmployeeId { get; set; }
+    [JsonProperty("employee_id")] 
+    public string EmployeeId { get; set; }
 
-    [JsonProperty("email")] public string Email { get; set; }
+    [JsonProperty("email")] 
+    public string Email { get; set; }
 
-    [JsonProperty("first_name")] public string FirstName { get; set; }
+    [JsonProperty("first_name")] 
+    public string FirstName { get; set; }
 
-    [JsonProperty("last_name")] public string LastName { get; set; }
+    [JsonProperty("last_name")] 
+    public string LastName { get; set; }
 
     [JsonProperty("dob")]
     public DateTime DateOfBirth { get; set; }
