+++ dotnet-test-samples/sqs-lambda/src/template.yaml
@@ -14,7 +14,7 @@ Resources:
     Type: AWS::Serverless::Function # More info about Function Resource: https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#awsserverlessfunction
     Properties:
       CodeUri: ./SqsEventHandler/
-      Handler: SqsEventHandler::SqsEventHandler.ProcessEmployeeFunction::Handler
+      Handler: SqsEventHandler::SqsEventHandler.Functions.ProcessEmployeeFunction::Handler
       Runtime: dotnet6
       Architectures:
         - x86_64
@@ -48,5 +48,3 @@ Outputs:
   ProcessEmployeeQueueUrl:
     Description: "Process Employee Queue Url"
     Value: !Ref ProcessEmployeeQueue
-
-
